requirements.txt
olvid-bot>=1.1.0
sphinx==7.4.7
sphinx-click==5.1.0
sphinx-panels==0.4.1
sphinx-rtd-theme==2.0.0
sphinx-design==0.6.1
myst-parser==4.0.0
sphinx-book-theme==1.1.3
sphinx-autobuild==2024.10.3
sphinx-intl==2.3.0
